split windows backend into multiple files
